Sort out intialization in processor
